+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/ModuleRegistry.kt
@@ -71,6 +71,18 @@ class ModuleRegistry(
     }
   }
 
+  /**
+   * Post onCreate event to all modules. It has its own method to ensure that it’s called first.
+   */
+  fun postOnCreate() {
+    forEach {
+      it.post(EventName.MODULE_CREATE)
+    }
+    registerActivityContracts()
+    readyForPostingEvents()
+    flushTheEventQueue()
+  }
+
   fun post(eventName: EventName) {
     if (addToQueueIfNeeded(eventName)) {
       return
@@ -117,11 +129,11 @@ class ModuleRegistry(
   /**
    * Tell the modules registry it can handle events as they come, without adding them to the event queue.
    */
-  fun readyForPostingEvents() = synchronized(this) {
+  private fun readyForPostingEvents() = synchronized(this) {
     isReadyForPostingEvents = true
   }
 
-  fun flushTheEventQueue() = synchronized(this) {
+  private fun flushTheEventQueue() = synchronized(this) {
     eventQueue.forEach { event ->
       forEach {
         event.post(it)